Fix signed/unsigned mismatch
git-svn-id: http://inethack.googlecode.com/svn/trunk@105 5aa2a760-743d-11de-9207-fb4211ab666b

+++ Classes/ShortcutView.h
@@ -28,7 +28,7 @@
 	
 	UIFont *font;
 	NSArray *shortcuts;
-	int currentIndex;
+	NSUInteger currentIndex;
 	int maxShortcutsOnScreen;
 	CGSize tileSize;
 	int recentlyTouchedItem;
